Handle the case of editing the model !!

+++ Model/Behavior/FileBehavior.php
@@ -12,6 +12,12 @@ class FileBehavior extends ModelBehavior {
 	}
 
 	public function beforeSave($model, $created) {
+		if ($model->data[$model->alias]['id']) {
+			// Updating, delete existing files.
+			$this->deleteAllFiles($model);
+		}
+
+		// Set afresh.
 		$cgi_data = $model->data[$model->alias]['cgi_data'];
 		$upload_folder = $this->getUploadFolder($model, $model->data[$model->alias]['filepath']);
 		$file = $upload_folder.$model->data[$model->alias]['filename'];
@@ -31,22 +37,19 @@ class FileBehavior extends ModelBehavior {
 	 * Delete all the existing files. Used mostly during an update.
 	 */
 	public function deleteAllFiles($model) {
-		$attachment = $model->find('first', array(
-			'conditions' => array(
-				'id' => $model->data['id']
-			),
-		));
-
-		$dir = $this->getUploadFolder($model, $attachment['type']);
+		$attachment = $model->findById($model->data[$model->alias]['id']);
 
 		//delete the original file
-		$this->deleteFile($dir . $attachment['filename']);
+		$this->deleteFile($attachment[$model->alias]['uri']);
 
 		//check if exists thumbs to be deleted too
-		$files = glob($dir . '*.' . $attachment['filename']);
-		if (is_array($files)) {
-			foreach ($files as $fileToDelete) {
-				$this->deleteFile($fileToDelete);
+		if (isset($attachment[$model->alias]['aspect_uri'])
+				&& !empty($attachment[$model->alias]['aspect_uri'])) {
+			$files = glob($attachment[$model->alias]['aspect_uri'].DS.'*');
+			if (is_array($files)) {
+				foreach ($files as $fileToDelete) {
+					$this->deleteFile($fileToDelete);
+				}
 			}
 		}
 	}
